UC Berkley CS61A Fall 2011 Final
The D33P language includes three types of tokens: open parentheses, close parentheses, and integers. 
An expression is well-formed if it contains balanced parentheses, and each integer correctly indicates its depth: the number of nested sets of parentheses that surround that integer.
Implement correct depth, which takes a list of tokens as input and returns True if and only if a prefix of the input is a well-formed D33P expression.

Assume that the input contains a balanced set of nested parentheses with single-digit positive integers surrounded by parentheses. 
You only need to check that the integers indicate the correct depths